# herecrm-pwa-openapi.yaml
openapi: 3.1.0
info:
  title: HereCRM PWA API
  version: 1.0.0
paths:
...
  /api/v1/pwa/chat/history/{customer_id}:
    get:
      tags:
      - pwa-chat
      summary: Get Chat History
      operationId: get_chat_history_api_v1_pwa_chat_history__customer_id__get
      parameters:
      - name: customer_id
        in: path
        required: true
        schema:
          type: integer
          title: Customer Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
                title: Response Get Chat History Api V1 Pwa Chat History  Customer
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/chat/send:
    post:
      tags:
      - pwa-chat
      summary: Send Message
      operationId: send_message_api_v1_pwa_chat_send_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSendRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/chat/execute:
    post:
      tags:
      - pwa-chat
      summary: Execute Tool
      operationId: execute_tool_api_v1_pwa_chat_execute_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatExecuteRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/pwa/chat/message/{message_id}:
    patch:
      tags:
      - pwa-chat
      summary: Update Message
      operationId: update_message_api_v1_pwa_chat_message__message_id__patch
      parameters:
      - name: message_id
        in: path
        required: true
        schema:
          type: integer
          title: Message Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ChatExecuteRequest:
      properties:
        tool_name:
          type: string
          maxLength: 100
          title: Tool Name
        arguments:
          additionalProperties: true
          type: object
          title: Arguments
      type: object
      required:
      - tool_name
      - arguments
      title: ChatExecuteRequest
    ChatMessage:
      properties:
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        role:
          $ref: '#/components/schemas/MessageRole'
        content:
          type: string
          maxLength: 5000
          title: Content
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        is_outbound:
          type: boolean
          title: Is Outbound
        is_executed:
          type: boolean
          title: Is Executed
          default: false
      type: object
      required:
      - role
      - content
      - timestamp
      - is_outbound
      title: ChatMessage
    ChatMessageUpdate:
      properties:
        message:
          type: string
          maxLength: 5000
          title: Message
      type: object
      required:
      - message
      title: ChatMessageUpdate
    ChatSendRequest:
      properties:
        customer_id:
          type: integer
          title: Customer Id
        message:
          type: string
          maxLength: 5000
          title: Message
        retry_id:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Retry Id
      type: object
      required:
      - customer_id
      - message
      title: ChatSendRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MessageRole:
      type: string
      enum:
      - USER
      - ASSISTANT
      title: MessageRole
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...